# krt: move event processing to queue