--- a/pilot/pkg/serviceregistry/kube/controller/ambient/waypoints.go
+++ b/pilot/pkg/serviceregistry/kube/controller/ambient/waypoints.go
@@ -207,7 +207,7 @@ func (a *index) WaypointsCollection(
 	gateways krt.Collection[*v1beta1.Gateway],
 	gatewayClasses krt.Collection[*v1beta1.GatewayClass],
 	pods krt.Collection[*v1.Pod],
-	withDebug krt.CollectionOption,
+	opts KrtOptions,
 ) krt.Collection[Waypoint] {
 	podsByNamespace := krt.NewNamespaceIndex(pods)
 	return krt.NewCollection(gateways, func(ctx krt.HandlerContext, gateway *v1beta1.Gateway) *Waypoint {
@@ -242,7 +242,7 @@ func (a *index) WaypointsCollection(
 		}
 
 		return a.makeWaypoint(gateway, gatewayClass, serviceAccounts, trafficType)
-	}, krt.WithName("Waypoints"), withDebug)
+	}, opts.WithName("Waypoints")...)
 }
 
 func makeInboundBinding(gateway *v1beta1.Gateway, gatewayClass *v1beta1.GatewayClass) *InboundBinding {