--- a/pkg/kube/kclient/client.go
+++ b/pkg/kube/kclient/client.go
@@ -176,6 +176,12 @@ func (n *informerClient[T]) ShutdownHandlers() {
 	}
 }
 
+type neverReady struct{}
+
+func (a neverReady) HasSynced() bool {
+	return false
+}
+
 func (n *informerClient[T]) AddEventHandler(h cache.ResourceEventHandler) cache.ResourceEventHandlerRegistration {
 	fh := cache.FilteringResourceEventHandler{
 		FilterFunc: func(obj interface{}) bool {
@@ -195,7 +201,7 @@ func (n *informerClient[T]) AddEventHandler(h cache.ResourceEventHandler) cache.
 	reg, err := n.informer.AddEventHandler(fh)
 	if err != nil {
 		// Should only happen if its already stopped. We should exit early.
-		return nil
+		return neverReady{}
 	}
 	n.registeredHandlers = append(n.registeredHandlers, handlerRegistration{registration: reg, handler: h})
 	return reg