… pilot/pkg/serviceregistry/kube/controller/ambient/ambientindex_serviceentry_test.go
… pilot/pkg/serviceregistry/kube/controller/ambient/ambientindex_test.go
… pilot/pkg/serviceregistry/kube/controller/ambient/ambientindex_workloadentry_test.go
… pilot/pkg/serviceregistry/kube/controller/ambient/authorization_test.go
… pilot/pkg/serviceregistry/kube/controller/ambient/workloads_test.go
… pkg/kube/krt/collection_test.go
… pkg/kube/krt/conformance_test.go
… pkg/kube/krt/core.go
… pkg/kube/krt/fetch.go
… pkg/kube/krt/index_test.go
… pkg/kube/krt/informer.go
… pkg/kube/krt/informer_test.go
… pkg/kube/krt/internal.go
… pkg/kube/krt/join_test.go
… pkg/kube/krt/leak_test.go
… pkg/kube/krt/processor.go
… pkg/kube/krt/recomputetrigger_test.go
… pkg/kube/krt/singleton_test.go
… pkg/kube/krt/static.go
… pkg/kube/krt/static_test.go
… pkg/kube/krt/sync.go
… pkg/kube/krt/testing.go
… pkg/queue/instance.go