tl;dr stop deadlocks in krt

Currently, KRT handlers are synchronous. This means that if I have a chain of N collections depending on each other, by the time we reach the Nth collection, our call stack will have the entire stack of N previous collections. Each other these will hold a lock while the handlers are called.

This has a few issues. The biggest issue is **it can deadlock**. I believe this is only possible with some specific sequences of dependencies; unfortunately, I do not fully understand the specific scenarios that trigger it to make it easily reproducible. However, I and others have encountered it on at least 4 occasions (with either changes we didn't end up pushing as-is). The root of this comes from the fact we have multiple entrypoints into recomputation (since we have many dependencies, and each will call our update handler). Additionally, when we do this recomputation, we need to register new handlers on new dependencies, which in turn requires a lock -- this is, as far as I know, the real root cause:
![image](https://github.com/user-attachments/assets/6d8a4008-8929-44c7-9f33-259fa2e8fad4)
Here we can see we may be holding locks to process an event from collection 1, then Fetch collection 1 which attempts to acquires a lock on collection 1: **deadlock**.

Note that while this may seem a contrived setup, its only really because its showing a simple case. Imagine I wanted to depend on `Waypoints` and `Pods` -- a seemingly reasonably use case and setup, but can trigger this.

Another plausible benefit of this is allowing krt to better make use of multiple cores, but I haven't verified this.

Benchmarks:
```
$ benchstat /tmp/old /tmp/new
name                   old time/op    new time/op    delta
Controllers/krt-16        107ms ±11%     100ms ± 5%   ~     (p=0.095 n=5+5)
Controllers/legacy-16    94.4ms ± 9%   100.5ms ± 7%   ~     (p=0.095 n=5+5)

name                   old alloc/op   new alloc/op   delta
Controllers/krt-16       63.4MB ± 0%    63.3MB ± 0%   ~     (p=0.222 n=5+5)
Controllers/legacy-16    60.8MB ± 0%    60.9MB ± 0%   ~     (p=0.730 n=4+5)

name                   old allocs/op  new allocs/op  delta
Controllers/krt-16         763k ± 0%      760k ± 0%   ~     (p=0.095 n=5+5)
Controllers/legacy-16      719k ± 0%      719k ± 0%   ~     (p=0.690 n=5+5)
```

Running all ambient tests:
```
3773 runs so far, 0 failures (100% pass rate). 1.355801297s avg, 3.75083996s max, 1.150561711s min
```

---

## Implementation comments

The core of this change is in the manyCollection implementation. We lift heavily from the Kubernetes informer model here, even pulling their processorListener almost directly. The primary change here is, of course, to implement a queue and to push all incoming events onto this queue for processing. However, a second major change is required too: implementing a distinct queue for each handler.

A few tests have been changed to not make assumptions about invalid ordering assumptions; these tests had a setup like 

```mermaid
graph TD
  a-->b
  c-->b
```

And assumed when `c` got an even it meant `a` did as well; this assumption no longer holds as each handler is independent.

Additionally, because of the async nature, computations may continue to occur after tests have complete but before things are shutdown. This causes race conditions when we mutate global state for feature flags. To handle this, the ambient code has been rewritten to avoid dependencies on mutable global state.