@howardjohn 2024-11-20
Tests are failing due to us not ever actually killing the stopCh (both with and without this change!) and that meaning we leak goroutines now but not before. Will work on a fix, it doesn't impact the main changes as its a bug in the usage of krt not krt itself.