@istio-testing 2024-11-20
@howardjohn: The following tests **failed**, say `/retest` to rerun all failed tests or `/retest-required` to rerun all mandatory failed tests:

Test name | Commit | Details | Required | Rerun command
--- | --- | --- | --- | ---
unit-tests-arm64_istio | 4bbcbf310afcba35f1cdf85ef5340d2b33f890ed | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/53994/unit-tests-arm64_istio/1859358831004356608) | true | `/test unit-tests-arm64`
unit-tests_istio | 4bbcbf310afcba35f1cdf85ef5340d2b33f890ed | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/53994/unit-tests_istio/1859358831830634496) | true | `/test unit-tests`
lint_istio | b9ee32b9b0a662f8df7a7cc86a33defccc22ab6a | [link](https://prow.istio.io/view/gs/istio-prow/pr-logs/pull/istio_istio/53994/lint_istio/1859365277725102080) | true | `/test lint`

<details>

Instructions for interacting with me using PR comments are available [here](https://git.k8s.io/community/contributors/guide/pull-requests.md).  If you have questions or suggestions related to my behavior, please file an issue against the [kubernetes-sigs/prow](https://github.com/kubernetes-sigs/prow/issues/new?title=Prow%20issue:) repository. I understand the commands that are listed [here](https://go.k8s.io/bot-commands).
</details>
<!-- test report -->